Add files via upload
Retail stores manage stocks everyday typically to sort the items in terms of deletion,
insertion and search operations. Stocks are frequently searched from the iventory by users
to count and check for available and unavailable items in store. They are also taken out
(deletion operstion) from the iventory to be distributed to dealers, customers or other
different branches. New items from suppliers can also be added in (insertion operation) by
users in the retail store iventory for storing purpose for future usage, though the insertion
will not be as frequent as the searching operation for the stocks in the iventory.